+++ crates/gitbutler-branch/src/base.rs
@@ -3,7 +3,7 @@ use std::{path::Path, time};
 use anyhow::{anyhow, Context, Result};
 use git2::Index;
 use gitbutler_branchstate::{VirtualBranchesAccess, VirtualBranchesHandle};
-use gitbutler_core::project_repository::RepoActions;
+use gitbutler_repo::{LogUntil, RepoActions};
 use serde::Serialize;
 
 use super::r#virtual as vb;
@@ -16,7 +16,7 @@ use gitbutler_core::virtual_branches::{branch, target, BranchId, GITBUTLER_INTEG
 use gitbutler_core::{error::Marker, git::RepositoryExt};
 use gitbutler_core::{
     git::{self, diff},
-    project_repository::{self, LogUntil},
+    project_repository,
     projects::FetchResult,
     virtual_branches::branch::BranchOwnershipClaims,
 };
@@ -588,7 +588,7 @@ pub fn target_to_base_branch(
 
     // gather a list of commits between oid and target.sha
     let upstream_commits = project_repository
-        .log(oid, project_repository::LogUntil::Commit(target.sha))
+        .log(oid, LogUntil::Commit(target.sha))
         .context("failed to get upstream commits")?
         .iter()
         .map(commit_to_remote_commit)
