+++ crates/gitbutler-branch/src/controller.rs
@@ -2,7 +2,7 @@ use anyhow::Result;
 use gitbutler_branchstate::{VirtualBranchesAccess, VirtualBranchesHandle};
 use gitbutler_core::{
     git::{credentials::Helper, BranchExt, RepositoryExt},
-    project_repository::{ProjectRepo, RepoActions},
+    project_repository::ProjectRepo,
     projects::FetchResult,
     types::ReferenceName,
 };
@@ -11,6 +11,7 @@ use gitbutler_oplog::{
     oplog::Oplog,
     snapshot::Snapshot,
 };
+use gitbutler_repo::RepoActions;
 use std::{path::Path, sync::Arc};
 
 use tokio::sync::Semaphore;
