+++ crates/gitbutler-branch/src/integration.rs
@@ -6,16 +6,14 @@ use bstr::ByteSlice;
 use gitbutler_branchstate::{VirtualBranchesAccess, VirtualBranchesHandle};
 use gitbutler_core::error::Marker;
 use gitbutler_core::git::RepositoryExt;
-use gitbutler_core::project_repository::RepoActions;
 use gitbutler_core::virtual_branches::{
     GITBUTLER_INTEGRATION_COMMIT_AUTHOR_EMAIL, GITBUTLER_INTEGRATION_COMMIT_AUTHOR_NAME,
     GITBUTLER_INTEGRATION_REFERENCE,
 };
 use gitbutler_core::{
-    git::CommitExt,
-    project_repository::{self, LogUntil},
-    virtual_branches::branch::BranchCreateRequest,
+    git::CommitExt, project_repository, virtual_branches::branch::BranchCreateRequest,
 };
+use gitbutler_repo::{LogUntil, RepoActions};
 
 use crate::conflicts;
