+++ crates/gitbutler-core/src/project_repository/mod.rs
@@ -2,4 +2,4 @@ mod config;
 mod repository;
 
 pub use config::Config;
-pub use repository::{LogUntil, ProjectRepo, RepoActions};
+pub use repository::ProjectRepo;
