+++ crates/gitbutler-repo/src/lib.rs
@@ -1,1 +1,4 @@
 pub mod rebase;
+
+mod repository;
+pub use repository::{LogUntil, RepoActions};
